+++ apps/mobile/src/features/ExecuteTx/components/ReviewAndExecute/ReviewAndExecuteContainer.tsx
@@ -18,7 +18,7 @@ import { selectActiveChain } from '@/src/store/chains'
 import { ExecutionMethod } from '@/src/features/HowToExecuteSheet/types'
 import { getExecutionMethod } from './helpers'
 import { parseFeeParams } from '@/src/utils/feeParams'
-import { useOptionalWalletConnectContext } from '@/src/features/WalletConnect/context/WalletConnectContext'
+import { useOptionalWalletConnectContext } from '@/src/features/WalletConnect/Signer/context/WalletConnectContext'
 import useGasFee from '../../hooks/useGasFee'
 import { useTransactionExecution } from '../../hooks/useTransactionExecution'
 import { useExecutionFunds } from '../../hooks/useExecutionFunds'
